feat: supports collection config
