+++ packages/plugin-stripe/demo/src/collections/Customers.ts
@@ -1,7 +1,4 @@
 import { CollectionConfig } from 'payload/types';
-import { createNewStripeCustomer } from '../hooks/createNewStripeCustomer';
-import { deleteStripeCustomer } from '../hooks/deleteStripeCustomer';
-import { syncExistingStripeCustomer } from '../hooks/syncExistingStripeCustomer';
 
 const Customers: CollectionConfig = {
   slug: 'customers',
@@ -15,42 +12,12 @@ const Customers: CollectionConfig = {
     ]
   },
   auth: true,
-  hooks: {
-    beforeValidate: [
-      createNewStripeCustomer,
-    ],
-    afterChange: [
-      syncExistingStripeCustomer,
-    ],
-    afterDelete: [
-      deleteStripeCustomer
-    ]
-  },
   fields: [
     {
       name: 'name',
       label: 'Name',
       type: 'text',
     },
-    {
-      name: 'stripeCustomerID',
-      label: 'Stripe Customer ID',
-      type: 'text',
-      saveToJWT: true,
-      admin: {
-        position: 'sidebar',
-        readOnly: true,
-      },
-    },
-    {
-      name: 'isSyncedToStripe',
-      label: 'Synced To Sync',
-      type: 'checkbox',
-      admin: {
-        position: 'sidebar',
-        readOnly: true,
-      },
-    }
   ]
 }
