--- packages/plugin-stripe/demo/src/mocks/serverModule.js
@@ -1,8 +1,0 @@
-export default {};
-export const syncExistingStripeCustomer = () => { }
-export const createNewStripeCustomer = () => { }
-export const deleteStripeCustomer = () => { }
-export const stripeREST = () => { }
-export const stripeWebhooks = () => { }
-export const stripeProxy = () => { }
-export const Stripe = () => { }
