+++ packages/plugin-stripe/demo/src/payload.config.ts
@@ -3,9 +3,7 @@ import path from 'path';
 import stripePlugin from '../../src';
 import Users from './collections/Users';
 import Customers from './collections/Customers';
-import { handleWebhooks } from './webhooks/handleWebhooks';
-
-const mockModulePath = path.resolve(__dirname, 'mocks/serverModule.js');
+// import { handleWebhooks } from '../../src/webhooks/handleWebhooks';
 
 export default buildConfig({
   serverURL: process.env.PAYLOAD_PUBLIC_CMS_URL,
@@ -19,10 +17,6 @@ export default buildConfig({
           alias: {
             ...config.resolve.alias,
             "payload": path.join(__dirname, "../node_modules/payload"),
-            "express": mockModulePath,
-            [path.resolve(__dirname, 'hooks/createNewStripeCustomer')]: mockModulePath,
-            [path.resolve(__dirname, 'hooks/syncExistingStripeCustomer')]: mockModulePath,
-            [path.resolve(__dirname, 'hooks/deleteStripeCustomer')]: mockModulePath,
           },
         },
       };
@@ -46,8 +40,9 @@ export default buildConfig({
   plugins: [
     stripePlugin({
       stripeSecretKey: process.env.STRIPE_SECRET_KEY,
+      collections: ['customers'],
       stripeWebhooksEndpointSecret: process.env.STRIPE_WEBHOOKS_ENDPOINT_SECRET,
-      webhooks: handleWebhooks
+      // webhooks: handleWebhooks, TODO: uncomment when custom root endpoints are released, then remove the custom route in server.ts
     }),
   ],
   typescript: {
