+++ packages/plugin-stripe/src/extendWebpackConfig.ts
@@ -18,17 +18,11 @@ export const extendWebpackConfig = (config: Config): ((webpackConfig: WebpackCon
         alias: {
           ...(existingWebpackConfig.resolve?.alias ? existingWebpackConfig.resolve.alias : {}),
           "stripe": mockModulePath,
-          // "fs": mockModulePath,
-          // "stream": mockModulePath,
-          // "url": mockModulePath,
-          // "util": mockModulePath,
-          // "querystring": mockModulePath,
-          // "net": mockModulePath,
-          // "zlib": mockModulePath,
-          // "async_hooks": mockModulePath,
-          // [path.resolve(__dirname, './index')]: mockModulePath,
           [path.resolve(__dirname, './routes/rest')]: mockModulePath,
           [path.resolve(__dirname, './routes/webhooks')]: mockModulePath,
+          [path.resolve(__dirname, './hooks/createNewStripeCustomer')]: mockModulePath,
+          [path.resolve(__dirname, './hooks/deleteStripeCustomer')]: mockModulePath,
+          [path.resolve(__dirname, './hooks/syncExistingStripeCustomer')]: mockModulePath,
         },
       },
     }
