+++ packages/plugin-stripe/src/hooks/deleteStripeCustomer.ts
@@ -1,8 +1,8 @@
-import type { CollectionAfterChangeHook, CollectionAfterDeleteHook, CollectionBeforeDeleteHook } from 'payload/types';
+import type { CollectionAfterDeleteHook, } from 'payload/types';
 import Stripe from 'stripe';
 
 const stripeSecretKey = process.env.STRIPE_SECRET_KEY;
-const stripe = new Stripe(stripeSecretKey, { apiVersion: '2022-08-01' });
+const stripe = new Stripe(stripeSecretKey || '', { apiVersion: '2022-08-01' });
 
 export const deleteStripeCustomer: CollectionAfterDeleteHook = async ({ req, doc }) => {
   const { payload } = req;
@@ -14,8 +14,8 @@ export const deleteStripeCustomer: CollectionAfterDeleteHook = async ({ req, doc
       await stripe.customers.del(doc.stripeCustomerID);
 
       console.log(`Successfully deleted customer ID: '${doc.stripeCustomerID}' from Stripe.`);
-    } catch (error) {
-      payload.logger.error(error.message);
+    } catch (error: any) {
+      payload.logger.error(`Error deleting customer ID: '${doc.stripeCustomerID}' from Stripe: ${error?.message || ''}`);
     }
   }
 }
