+++ packages/plugin-stripe/src/index.ts
@@ -4,8 +4,13 @@ import { stripeREST } from './routes/rest';
 // import express from 'express';
 import { StripeConfig } from './types';
 import { extendWebpackConfig } from './extendWebpackConfig';
+import { createNewStripeCustomer } from './hooks/createNewStripeCustomer';
+import { syncExistingStripeCustomer } from './hooks/syncExistingStripeCustomer';
+import { deleteStripeCustomer } from './hooks/deleteStripeCustomer';
 
 const stripePlugin = (stripeConfig: StripeConfig) => (config: Config): Config => {
+  const { collections } = config;
+
   return ({
     ...config,
     admin: {
@@ -32,7 +37,61 @@ const stripePlugin = (stripeConfig: StripeConfig) => (config: Config): Config =>
           stripeREST(req, res, next, stripeConfig)
         }
       },
-    ]
+    ],
+    collections: collections?.map((collection) => {
+      const {
+        hooks: existingHooks
+      } = collection;
+
+      const enabledCollections = stripeConfig.collections || [];
+
+      const isEnabled = enabledCollections.indexOf(collection.slug) > -1;
+
+      if (isEnabled) {
+        return {
+          ...collection,
+          hooks: {
+            ...collection.hooks,
+            beforeValidate: [
+              ...(existingHooks?.beforeValidate || []),
+              createNewStripeCustomer
+            ],
+            afterChange: [
+              ...(existingHooks?.afterChange || []),
+              syncExistingStripeCustomer
+            ],
+            afterDelete: [
+              ...(existingHooks?.afterDelete || []),
+              deleteStripeCustomer
+            ],
+          },
+          fields: [
+            ...collection.fields,
+            {
+              name: 'stripeCustomerID',
+              label: 'Stripe Customer ID',
+              type: 'text',
+              saveToJWT: true,
+              admin: {
+                position: 'sidebar',
+                readOnly: true,
+              },
+            },
+            {
+              name: 'isSyncedToStripe',
+              label: 'Synced To Sync',
+              type: 'checkbox',
+              admin: {
+                position: 'sidebar',
+                readOnly: true,
+              },
+            }
+          ]
+        };
+      }
+
+      return collection;
+    })
   })
 };
