+++ packages/plugin-stripe/src/types.ts
@@ -1,6 +1,7 @@
-import Stripe from "stripe"
+import { Payload } from "payload";
+import Stripe from "stripe";
 
-export type StripeWebhookHandler = (event: any, stripe: Stripe, stripeConfig?: StripeConfig) => void;
+export type StripeWebhookHandler = (payload: Payload, event: any, stripe: Stripe, stripeConfig?: StripeConfig) => void;
 
 export type StripeWebhookHandlers = {
   [webhookName: string]: StripeWebhookHandler
@@ -10,6 +11,7 @@ export type StripeConfig = {
   stripeSecretKey: string
   stripeWebhooksEndpointSecret?: string
   webhooks?: StripeWebhookHandler | StripeWebhookHandlers
+  collections?: string[]
 }
 
 export type StripeProxy = (args: {
